+++ AlternativeTextures/Framework/UI/SprayCanMenu.cs
@@ -59,11 +59,11 @@ namespace AlternativeTextures.Framework.UI
                 myID = 8997
             };
 
-            this.allButton = new ClickableComponent(new Rectangle((int)_sideBarPosition.X + 23, (int)radiusAddButton.bounds.Y + 125, 100, 45), "allButton")
+            this.allButton = new ClickableComponent(new Rectangle((int)_sideBarPosition.X + 23, radiusAddButton.bounds.Y + 125, 100, 45), "allButton")
             {
                 myID = 8991
             };
-            this.noneButton = new ClickableComponent(new Rectangle((int)_sideBarPosition.X + 23, (int)radiusAddButton.bounds.Y + 175, 100, 45), "noneButton")
+            this.noneButton = new ClickableComponent(new Rectangle((int)_sideBarPosition.X + 23, radiusAddButton.bounds.Y + 175, 100, 45), "noneButton")
             {
                 myID = 8992
             };
@@ -196,7 +196,7 @@ namespace AlternativeTextures.Framework.UI
             Utility.drawTextWithShadow(b, _tileRadius.ToString(), Game1.smallFont, new Vector2(radiusSubtractButton.bounds.X + 48, radiusSubtractButton.bounds.Y), Game1.textColor);
             radiusAddButton.draw(b);
 
-            Utility.drawTextWithShadow(b, "  ---\nQuick\nSelect", Game1.smallFont, new Vector2(_sideBarPosition.X + 38, (int)radiusAddButton.bounds.Y + 30), Game1.textColor);
+            Utility.drawTextWithShadow(b, "  ---\nQuick\nSelect", Game1.smallFont, new Vector2(_sideBarPosition.X + 38, radiusAddButton.bounds.Y + 30), Game1.textColor);
             IClickableMenu.drawTextureBox(b, Game1.mouseCursors, new Rectangle(432, 439, 9, 9), allButton.bounds.X, allButton.bounds.Y, allButton.bounds.Width, allButton.bounds.Height, Color.White, 4f, drawShadow: false);
             Utility.drawTextWithShadow(b, "All", Game1.smallFont, new Vector2(allButton.bounds.X + 35, allButton.bounds.Y + 5), Game1.textColor);
             IClickableMenu.drawTextureBox(b, Game1.mouseCursors, new Rectangle(432, 439, 9, 9), noneButton.bounds.X, noneButton.bounds.Y, noneButton.bounds.Width, noneButton.bounds.Height, Color.White, 4f, drawShadow: false);
@@ -204,9 +204,9 @@ namespace AlternativeTextures.Framework.UI
 
             var enabledTextureSum = _selectedIdsToModels.Sum(t => t.Value.Variations.Count).ToString();
             var numberOfDigitsInSum = enabledTextureSum.Length;
-            Utility.drawTextWithShadow(b, "  ---", Game1.smallFont, new Vector2(_sideBarPosition.X + 38, (int)noneButton.bounds.Y + 60), Game1.textColor);
-            Utility.drawTextWithShadow(b, "Selected\nTextures", Game1.smallFont, new Vector2(_sideBarPosition.X + 25, (int)noneButton.bounds.Y + 85), Game1.textColor);
-            Utility.drawTextWithShadow(b, enabledTextureSum, Game1.smallFont, new Vector2(_sideBarPosition.X + 65 - (enabledTextureSum.Length >= 2 ? 5 * (enabledTextureSum.Length - 1) : 0), (int)noneButton.bounds.Y + 145), Game1.textColor);
+            Utility.drawTextWithShadow(b, "  ---", Game1.smallFont, new Vector2(_sideBarPosition.X + 38, noneButton.bounds.Y + 60), Game1.textColor);
+            Utility.drawTextWithShadow(b, "Selected\nTextures", Game1.smallFont, new Vector2(_sideBarPosition.X + 25, noneButton.bounds.Y + 85), Game1.textColor);
+            Utility.drawTextWithShadow(b, enabledTextureSum, Game1.smallFont, new Vector2(_sideBarPosition.X + 65 - (enabledTextureSum.Length >= 2 ? 5 * (enabledTextureSum.Length - 1) : 0), noneButton.bounds.Y + 145), Game1.textColor);
 
             Game1.mouseCursorTransparency = 1f;
             base.drawMouse(b);
